+++ dz6/src/main/resources/application.yaml
@@ -1,0 +1,9 @@
+spring:
+   h2:
+      console:
+         enabled: true
+   jpa:
+      hibernate:
+         ddl-auto: none
+   liquibase:
+      change-log: classpath:db/liquibase-changelog.xml
