+++ dz6/src/main/resources/db/changelog/01-create-book-and-author-and-genre-schema.xml
@@ -1,0 +1,50 @@
+<databaseChangeLog
+	xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+         http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.6.xsd">
+
+	<changeSet id="01" author="me">
+		<createTable tableName="AUTHOR"
+			remarks="A table to contain all books">
+			<column name="ID" type="int" autoIncrement="true">
+				<constraints unique="true"
+					primaryKey="true" />
+					
+			</column>
+			<column name="NAME" type="varchar(255)">
+				<constraints nullable="false" unique="false" />
+			</column>
+			<column name="BIRTHDATE" type="DATE">
+				<constraints nullable="false" />
+			</column>
+		</createTable>
+
+		<createTable tableName="GENRE"
+			remarks="A table to contain all the genre">
+			<column name="ID" type="int" autoIncrement="true">
+				<constraints primaryKey="true" />
+			</column>
+			<column name="DESCRIPTION" type="varchar(100)">
+				<constraints nullable="false" />
+			</column>
+		</createTable>
+
+		<createTable tableName="BOOK"
+			remarks="A table to contain all the genre">
+			<column name="ID" type="int" autoIncrement="true">
+				<constraints primaryKey="true" />
+			</column>
+			<column name="BOOKNAME" type="varchar(100)">
+				<constraints nullable="false" />
+			</column>
+			<column name="AUTHORID" type="int">
+				<constraints nullable="false" primaryKey="true" />
+			</column>
+			<column name="GENREID" type="int">
+				<constraints nullable="false" primaryKey="true" />
+			</column>
+		</createTable>
+	</changeSet>
+	 
+</databaseChangeLog>
