+++ dz6/src/main/resources/db/liquibase-changelog.xml
@@ -1,0 +1,8 @@
+<databaseChangeLog
+  xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+         http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
+
+  <include file="changelog/01-create-book-and-author-and-genre-schema.xml" relativeToChangelogFile="true"/>
+</databaseChangeLog>
