+++ dz6/src/test/java/ru/akozlovskiy/springdz06/domain/TestApplication.java
@@ -2,7 +2,7 @@ package ru.akozlovskiy.springdz06.domain;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication 
+@SpringBootApplication
 public class TestApplication {
 
 }
