+++ dz6/src/test/java/ru/akozlovskiy/springdz06/domain/dao/jpa/AuthorDAOJpaTest.java
@@ -16,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import ru.akozlovskiy.springdz06.domain.Author;
@@ -25,6 +26,7 @@ import ru.akozlovskiy.springdz06.exception.DaoException;
 @DataJpaTest
 @Import({ AuthorDAOJpa.class })
 @DisplayName("DAO севрис по работе с авторами")
+@Sql(scripts = "classpath:db/testdata.sql")
 public class AuthorDAOJpaTest {
 
 	private static final String AUTHOR_BIRTH_DATE = "1891-05-15";
