+++ dz6/src/test/java/ru/akozlovskiy/springdz06/domain/dao/jpa/BookDAOJpaTest.java
@@ -2,6 +2,8 @@ package ru.akozlovskiy.springdz06.domain.dao.jpa;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
 
@@ -10,7 +12,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import ru.akozlovskiy.springdz06.domain.Book;
@@ -20,6 +24,7 @@ import ru.akozlovskiy.springdz06.exception.DaoException;
 @DataJpaTest
 @Import({ BookDAOJpa.class, AuthorDAOJpa.class, GenreDAOJpa.class })
 @DisplayName("DAO сервис по работе с книгами")
+@Sql(scripts = "classpath:db/testdata.sql")
 public class BookDAOJpaTest {
 
 	private static final String BOOK_NAME_BD = "BOOK_NAME1";
@@ -28,6 +33,9 @@ public class BookDAOJpaTest {
 
 	@Autowired
 	private BookDAOJpa bookDAO;
+	
+	@Autowired
+	private TestEntityManager em;
 
 	@Test
 	@DisplayName("Поиск по ID")
@@ -64,5 +72,34 @@ public class BookDAOJpaTest {
 		assertEquals(book.getBookName(), BOOK_NAME_BD);
 		assertEquals(book.getAuthor().getName(), AUTHOR_NAME_IN_BD);
 		assertEquals(book.getGenre().getDescription(), GENRE_IN_BD);
+	}
+	
+	@Test
+	@DisplayName("Добавление в случае корректных данных")
+	public void testAddAndGetById() throws DaoException {
+		String testBookName = "booknametest";
+		long bookId = bookDAO.add(testBookName, AUTHOR_NAME_IN_BD, GENRE_IN_BD);
+		assertNotEquals(0, bookId);
+
+		Book book = em.find(Book.class, bookId);
+		assertEquals(book.getBookName(), testBookName);
+		assertEquals(book.getAuthor().getName(), AUTHOR_NAME_IN_BD);
+		assertEquals(book.getGenre().getDescription(), GENRE_IN_BD);
+	}
+
+	@Test
+	@DisplayName("Возврат ошибки при добавлении с не заведенным автором")
+	public void tesAddDaoWithWrongAuthor() throws DaoException {
+		assertThrows(DaoException.class, () -> {
+			bookDAO.add("bookname", "WRONG_AUTHOR_NAME", GENRE_IN_BD);
+		});
+	}
+
+	@Test
+	@DisplayName("Возврат ошибки при добавлении с не заведенным жанром")
+	public void tesAddDaoWithWrongGenre() throws DaoException {
+		assertThrows(DaoException.class, () -> {
+			bookDAO.add("bookname", AUTHOR_NAME_IN_BD, "WRONG_GENRE");
+		});
 	}
 }
