--- dz6/src/test/resources/application.yml
@@ -1,4 +1,0 @@
-spring:
-   jpa:
-      hibernate:
-         ddl-auto: none
